Repository を配置変更
