+++ src/main/java/x/plantree/infrastructure/NodeRepositoryImpl.java
@@ -1,4 +1,4 @@
-package x.plantree.repositories;
+package x.plantree.infrastructure;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -10,16 +10,18 @@ import org.springframework.jdbc.support.rowset.SqlRowSet;
 import org.springframework.stereotype.Repository;
 
 import x.plantree.domain.entities.Node;
+import x.plantree.domain.repositories.NodeRepository;
 
 @Repository
-public class NodeRepository {
+public class NodeRepositoryImpl implements NodeRepository {
 
   @Autowired
-  JdbcTemplate jdbcTemplate;
+  private JdbcTemplate jdbcTemplate;
 
   /**
    * データベースに Node を登録する
    */
+  @Override
   public Node save(Node node) {
     String sql = "INSERT INTO nodes VALUES (?, ?)";
     jdbcTemplate.update(sql, node.getId(), node.getTitle());
@@ -30,6 +32,7 @@ public class NodeRepository {
   /**
    * データベースから全ての Node を取得する
    */
+  @Override
   public List<Node> findAll() {
     String sql = "SELECT * FROM nodes";
     SqlRowSet rs = jdbcTemplate.queryForRowSet(sql);
@@ -44,6 +47,7 @@ public class NodeRepository {
   /**
    * データベースから特定の Node を取得する
    */
+  @Override
   public Node findById(int id) {
     String sql = "SELECT * FROM nodes WHERE id = ?";
     Map<String, Object> map = jdbcTemplate.queryForMap(sql, id);
@@ -56,6 +60,7 @@ public class NodeRepository {
   /**
    * データベースの特定の Node を更新する
    */
+  @Override
   public Node updateById(int id, Node node) {
     String sql = "UPDATE nodes SET title = ? WHERE id = ?";
     jdbcTemplate.update(sql, node.getTitle(), id);
@@ -65,6 +70,7 @@ public class NodeRepository {
   /**
    * データベースから特定の Node を削除する
    */
+  @Override
   public void deleteById(int id) {
     String sql = "DELETE FROM nodes WHERE id = ?";
     jdbcTemplate.update(sql, id);
