+++ src/main/java/x/plantree/services/NodeService.java
@@ -2,16 +2,36 @@ package x.plantree.services;
 
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
 import x.plantree.domain.entities.Node;
+import x.plantree.domain.repositories.NodeRepository;
 
-public interface NodeService {
-  public Node saveNode(Node node);
+@Service
+public class NodeService {
 
-  public List<Node> getNodeList();
+  @Autowired
+  private NodeRepository nodeRepository;
 
-  public Node getNodeById(int id);
+  public Node saveNode(Node node) {
+    return nodeRepository.save(node);
+  }
 
-  public Node updateNode(int id, Node node);
+  public List<Node> getNodeList() {
+    return nodeRepository.findAll();
+  }
 
-  public void deleteNode(int id);
+  public Node getNodeById(int id) {
+    return nodeRepository.findById(id);
+  }
+
+  public Node updateNode(int id, Node node) {
+    return nodeRepository.updateById(id, node);
+  }
+
+  public void deleteNode(int id) {
+    nodeRepository.deleteById(id);
+  }
+
 }
