+++ src/main/java/x/plantree/services/NodeServiceImplWithoutRepo.java
@@ -11,7 +11,7 @@ import x.plantree.errors.NotFoundException;
 import x.plantree.domain.entities.Node;
 
 @Service
-public class NodeServiceImplWithoutRepo implements NodeService {
+public class NodeServiceImplWithoutRepo implements NodeService_bk {
 
   private final AtomicInteger counter = new AtomicInteger();
